Extract the comment manager.
No functional change.
