+++ src/compiler/toitdoc_parser.cc
@@ -15,6 +15,7 @@
 
 #include <string>
 
+#include "comments.h"
 #include "scanner.h"
 #include "parser.h"
 #include "toitdoc_parser.h"
@@ -308,45 +309,24 @@ class ToitdocParser {
 
 /// Manages all existing comments, making it easier to find
 ///   toitdocs, and associating them with their respective AST nodes.
-class CommentsManager {
+class ToitdocCommentsManager : public CommentsManager {
  public:
-  CommentsManager(List<Scanner::Comment> comments,
-                  Source* source,
-                  SymbolCanonicalizer* symbols,
-                  Diagnostics* diagnostics)
-      : comments_(comments)
-      , source_(source)
+  ToitdocCommentsManager(List<Scanner::Comment> comments,
+                         Source* source,
+                         SymbolCanonicalizer* symbols,
+                         Diagnostics* diagnostics)
+      : CommentsManager(comments, source)
       , symbols_(symbols)
       , diagnostics_(diagnostics) {
     ASSERT(is_sorted(comments));
   }
 
-  int find_closest_before(ast::Node* node);
   Toitdoc<ast::Node*> find_for(ast::Node* node);
-  bool is_attached(int index1, int index2) {
-    return is_attached(comments_[index1].range(), comments_[index2].range(), false);
-  }
-  bool is_attached(Source::Range previous,
-                   Source::Range next,
-                   bool allow_modifiers);
   Toitdoc<ast::Node*> make_ast_toitdoc(int index);
 
  private:
-  List<Scanner::Comment> comments_;
-  Source* source_;
   SymbolCanonicalizer* symbols_;
   Diagnostics* diagnostics_;
-
-  int last_index_ = 0;
-
-  static bool is_sorted(List<Scanner::Comment> comments) {
-    for (int i = 1; i < comments.length(); i++) {
-      if (!comments[i - 1].range().from().is_before(comments[i].range().from())) {
-        return false;
-      }
-    }
-    return true;
-  }
 };
 
 int ToitdocSource::source_offset_at(int offset) const  {
@@ -1151,59 +1131,7 @@ void ToitdocParser::report_error(Source::Range range, const char* message) {
   diagnostics_->report_error(range, message);
 }
 
-int CommentsManager::find_closest_before(ast::Node* node) {
-  auto node_range = node->full_range();
-  if (node_range.is_before(comments_[0].range())) return -1;
-  if (comments_.last().range().is_before(node_range)) return comments_.length() - 1;
-
-  if (comments_[last_index_].range().is_before(node_range) &&
-      node_range.is_before(comments_[last_index_ + 1].range())) {
-    return last_index_;
-  }
-  int start = 0;
-  int end = comments_.length() - 1;
-  while (start < end) {
-    int mid = start + (end - start) / 2;
-    if (comments_[mid].range().is_before(node_range)) {
-      if (node_range.is_before(comments_[mid + 1].range())) {
-        return mid;
-      }
-      start = mid + 1;
-    } else {
-      end = mid;
-    }
-  }
-  return -1;
-}
-
-/// When [allow_modifiers] is true, allows modifiers on the line of the
-///   [next] range.
-bool CommentsManager::is_attached(Source::Range previous,
-                                  Source::Range next,
-                                  bool allow_modifiers) {
-  // Check that there is one newline, and otherwise only whitespace.
-  int start_offset = source_->offset_in_source(previous.to());
-  int end_offset = source_->offset_in_source(next.from());
-  int i = start_offset;
-  auto text = source_->text();
-  while (i < end_offset and text[i] == ' ') i++;
-  if (i == end_offset) return true;
-  if (text[i] == '\r') i++;
-  if (i == end_offset) return true;
-  if (text[i++] != '\n') return false;
-  while (i < end_offset and text[i] == ' ') i++;
-  if (i == end_offset) return true;
-  if (!allow_modifiers) return false;
-  for (; i < end_offset; i++) {
-    if (text[i] == '\n') return false;
-    if (text[i] == '\r') return false;
-    if (text[i] == ':') return false;
-  }
-  return true;
-}
-
-
-Toitdoc<ast::Node*> CommentsManager::find_for(ast::Node* node) {
+Toitdoc<ast::Node*> ToitdocCommentsManager::find_for(ast::Node* node) {
   auto not_found = Toitdoc<ast::Node*>::invalid();
   int closest = find_closest_before(node);
   if (closest == -1) return not_found;
@@ -1228,7 +1156,7 @@ Toitdoc<ast::Node*> CommentsManager::find_for(ast::Node* node) {
   return make_ast_toitdoc(closest_toit);
 }
 
-Toitdoc<ast::Node*> CommentsManager::make_ast_toitdoc(int index) {
+Toitdoc<ast::Node*> ToitdocCommentsManager::make_ast_toitdoc(int index) {
   // If the comment is a single line '///' comment, search for comments that
   // precede and succeed it.
   int first_toit = index;
@@ -1267,7 +1195,7 @@ void attach_toitdoc(ast::Unit* unit,
                     Diagnostics* diagnostics) {
   if (scanner_comments.is_empty()) return;
   ToitdocDiagnostics toitdoc_diagnostics(diagnostics);
-  CommentsManager comments_manager(scanner_comments, source, symbols, &toitdoc_diagnostics);
+  ToitdocCommentsManager comments_manager(scanner_comments, source, symbols, &toitdoc_diagnostics);
 
   ast::Node* earliest_declaration = null;
   for (auto declaration : unit->declarations()) {
